bang

+++ Hw5/functions.cpp
@@ -96,10 +96,9 @@ void output_image(const std::string& file,
     word_file << "P3\n" << width << " " << height << "\n" << "255" << "\n";
     for (int i = 0; i < height; i++){ // Possible problem in not switching rows and columns 
         for (int j = 0; j < width; j++){ //SHOULD BE TRANSPOSED NOW
-            word_file << pict[j][i].red << " " << pict[j][i].green << " " << pict[j][i].blue << "\n";
+            word_file << (int)pict[j][i].red << " " << (int)pict[j][i].green << " " << (int)pict[j][i].blue << "\n";
         }
-    }
-    // TODO(student): implement writing image to file
+    }    // TODO(student): implement writing image to file
 }
 
 double map_coordinates(size_t s_dim,
